Revert "Add challenges recyclerview"
This reverts commit 4d8076ec3c7e5fd3cb492736ac1ee40666919192.
