+++ app/src/main/java/com/wildcard/eMission/MainActivity.kt
@@ -15,20 +15,19 @@ class MainActivity : AppCompatActivity() {
         setContentView(R.layout.activity_main)
         val navView: BottomNavigationView = findViewById(R.id.nav_view)
         setSupportActionBar(findViewById(R.id.toolbar))
-        supportActionBar?.title = ""
 
         val navController = findNavController(R.id.nav_host_fragment)
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
-//        val appBarConfiguration = AppBarConfiguration(
-//            setOf(
-//                R.id.navigation_home,
-//                R.id.navigation_rewards,
-//                R.id.navigation_learning,
-//                R.id.navigation_you
-//            )
-//        )
-//        setupActionBarWithNavController(navController, appBarConfiguration)
+        val appBarConfiguration = AppBarConfiguration(
+            setOf(
+                R.id.navigation_home,
+                R.id.navigation_rewards,
+                R.id.navigation_learning,
+                R.id.navigation_you
+            )
+        )
+        setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
     }
 }
