+++ app/src/main/java/com/wildcard/eMission/YouFragment.kt
@@ -1,4 +1,4 @@
-package com.wildcard.eMission.ui.rewards
+package com.wildcard.eMission
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,22 +8,21 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
-import com.wildcard.eMission.R
 
-class RewardsFragment : Fragment() {
+class YouFragment : Fragment() {
 
-    private lateinit var rewardsViewModel: RewardsViewModel
+    private lateinit var youViewModel: YouViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        rewardsViewModel =
-            ViewModelProviders.of(this).get(RewardsViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_rewards, container, false)
-        val textView: TextView = root.findViewById(R.id.text_rewards)
-        rewardsViewModel.text.observe(this, Observer {
+        youViewModel =
+            ViewModelProviders.of(this).get(YouViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_you, container, false)
+        val textView: TextView = root.findViewById(R.id.text_you)
+        youViewModel.text.observe(this, Observer {
             textView.text = it
         })
         return root
