+++ app/src/main/java/com/wildcard/eMission/YouViewModel.kt
@@ -1,4 +1,4 @@
-package com.wildcard.eMission.ui.you
+package com.wildcard.eMission
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -7,7 +7,7 @@ import androidx.lifecycle.ViewModel
 class YouViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is You Fragment"
+        value = "This is you Fragment"
     }
     val text: LiveData<String> = _text
 }
