--- app/src/main/java/com/wildcard/eMission/model/Challenge.kt
@@ -1,17 +1,0 @@
-package com.wildcard.eMission.model
-
-data class Challenge (
-    val cId: String,
-    val name: String,
-    val description: String,
-    val points: Int,
-    var status: CompleteStatus,
-    val info: Info?,
-    val matchingLifestyle: Lifestyle
-)
-
-enum class CompleteStatus {
-    COMPLETE,
-    ONGOING,
-    UNSTARTED
-}
