--- app/src/main/java/com/wildcard/eMission/model/Reward.kt
@@ -1,14 +1,0 @@
-package com.wildcard.eMission.model
-
-data class Reward (
-    val rId: String,
-    val name: String,
-    val type: RewardType,
-    val content: Any?
-)
-
-enum class RewardType {
-    TITLE,
-    CHALLENGE_PACK,
-    ACTION
-}
