+++ app/src/main/java/com/wildcard/eMission/ui/learning/LearningFragment.kt
@@ -5,13 +5,10 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import androidx.appcompat.widget.Toolbar
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
-import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.wildcard.eMission.R
-import com.wildcard.eMission.Utils
 
 class LearningFragment : Fragment() {
 
@@ -30,28 +27,5 @@ class LearningFragment : Fragment() {
             textView.text = it
         })
         return root
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        setupActionBar()
-        activity?.findViewById<BottomNavigationView>(R.id.nav_view)?.itemTextColor = context?.getColorStateList(R.color.nav_item_color_state_list_3)
-        activity?.findViewById<BottomNavigationView>(R.id.nav_view)?.itemIconTintList = context?.getColorStateList(R.color.nav_item_color_state_list_3)
-    }
-
-    private fun setupActionBar() {
-        val actionBar = activity?.findViewById<Toolbar>(R.id.toolbar)
-
-        val actionBarTitle = actionBar?.findViewById<TextView>(R.id.actionbar_title)
-        val actionBarSubtitle = actionBar?.findViewById<TextView>(R.id.actionbar_subtitle)
-        actionBarTitle?.text = getString(R.string.title_learning)
-        actionBarSubtitle?.text = getString(R.string.subtitle_learning)
-
-        Utils.setGradientTextColor(
-            actionBarTitle!!,
-            context!!.getColor(R.color.colorPrimary_green),
-            context!!.getColor(R.color.colorPrimary_yellow)
-        )
     }
 }
