--- app/src/main/res/color/nav_item_color_state_list_1.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-<item android:state_selected="false" android:color="@color/lightGray" />
-<item android:color="@color/colorPrimary_blue" />
-</selector>
