--- app/src/main/res/drawable-anydpi/ic_challenge_checkpoint_filled.xml
@@ -1,14 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="20"
-    android:viewportHeight="20"
-    android:tint="#333333"
-    android:alpha="1">
-  <group android:translateX="-2"
-      android:translateY="-1.5833334">
-      <path
-          android:fillColor="#FF000000"
-          android:pathData="M12,17.27L18.18,21l-1.64,-7.03L22,9.24l-7.19,-0.61L12,2 9.19,8.63 2,9.24l5.46,4.73L5.82,21z"/>
-  </group>
-</vector>
