--- app/src/main/res/drawable-anydpi/ic_info.xml
@@ -1,14 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="20"
-    android:viewportHeight="20"
-    android:tint="#333333"
-    android:alpha="1">
-  <group android:translateX="-2"
-      android:translateY="-2">
-      <path
-          android:fillColor="#FF000000"
-          android:pathData="M11,17h2v-6h-2v6zM12,2C6.48,2 2,6.48 2,12s4.48,10 10,10 10,-4.48 10,-10S17.52,2 12,2zM12,20c-4.41,0 -8,-3.59 -8,-8s3.59,-8 8,-8 8,3.59 8,8 -3.59,8 -8,8zM11,9h2L13,7h-2v2z"/>
-  </group>
-</vector>
