--- app/src/main/res/drawable/gradient_progress_bar.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item android:id="@android:id/background">
-        <shape>
-            <corners android:radius="10dp"/>
-            <gradient
-                android:startColor="#4D2F80ED"
-                android:endColor="#4D01A39D"
-                />
-        </shape>
-    </item>
-
-    <item android:id="@android:id/progress">
-        <clip>
-            <shape>
-                <corners android:radius="10dp"/>
-                <gradient
-                    android:startColor="@color/colorPrimary_blue"
-                    android:endColor="@color/colorPrimary_green"
-                    />
-            </shape>
-        </clip>
-    </item>
-
-</layer-list>
