+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -8,12 +8,12 @@
 
     <item
         android:id="@+id/navigation_rewards"
-        android:icon="@drawable/ic_challenge_checkpoint_filled"
+        android:icon="@drawable/ic_dashboard_black_24dp"
         android:title="@string/title_rewards" />
 
     <item
         android:id="@+id/navigation_learning"
-        android:icon="@drawable/ic_info"
+        android:icon="@drawable/ic_notifications_black_24dp"
         android:title="@string/title_learning" />
     <item
         android:id="@+id/navigation_you"
