+++ app/src/main/res/values/preloaded_fonts.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <array name="preloaded_fonts" translatable="false">
-        <item>@font/lato</item>
         <item>@font/quicksand_medium</item>
     </array>
 </resources>
