+++ app/src/main/res/values/styles.xml
@@ -6,21 +6,6 @@
         <item name="colorPrimary">@color/colorPrimary</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
-        <item name="android:statusBarColor">#fff</item>
-        <item name="android:windowLightStatusBar">true</item>
-    </style>
-
-    <style name="ActionBarTitleAppearance" parent="Widget.AppCompat.ActionBar.TabText">
-        <item name="android:textSize">35sp</item>
-        <item name="android:fontFamily">@font/quicksand_medium</item>
-        <item name="android:textStyle">bold</item>
-    </style>
-
-    <style name="TitleAppearance" parent="TextAppearance.MaterialComponents.Headline6">
-        <item name="android:textSize">20sp</item>
-        <item name="android:fontFamily">@font/quicksand_medium</item>
-        <item name="android:textStyle">bold</item>
-        <item name="android:textColor">@color/textColor</item>
     </style>
 
 </resources>
